requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.3.25
langchain-core==0.3.72
langchain-openai==0.3.25
langchain-community==0.3.23
langgraph==0.6.0
langsmith==0.3.45
pypdf==3.17.4
chromadb==0.4.18
python-multipart==0.0.6
pydantic==2.10.5
pydantic-settings==2.8.0
python-dotenv==1.0.0
tavily-python==0.7.6
tiktoken==0.8
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2 